Add note about activating the debug mode
I was testing your plugin locally and rand into a situation where the debug toolbar was not shown but an error message instead even though I set debug to true. Took me quite some time untill I found out I can (and should) manually activate the debug mode for my IP trhough the CLI.

I'm not sure if this is really the best solution. Either it needs more documentation or it should respect the current setting for the debug mode.